pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Governed workflow CLI for the Inspected application"
authors = [
    {name = "Justin Cook", email = "justin@example.com"},
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "google-genai>=0.1.0",
    "PyYAML>=6.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
